corregido

+++ BackEnd/VistasU/Eliminar/Eliminar_Solicitud.php
@@ -23,7 +23,10 @@
 		$NombreCarpeta = $result['Usuario'];
 		$NombreDocumento = $result['Documento'];
 		
-		unlink('../Documentos/Solicitudes/'.$NombreCarpeta.'/'.$NombreDocumento);
+		if(isset($NombreDocumento)){
+			unlink('../Documentos/Solicitudes/'.$NombreCarpeta.'/'.$NombreDocumento);
+		}
+		
 
 		$sql = "DELETE FROM solicitudes WHERE IdSolicitudes=:IdSolicitudes";
 		$stmt = $pdo->prepare($sql);
